HACK: OMAPDSS: AM43x:Add FIFO size limitation in dss_feats
Due to bandwidth constraits seen on the AM43x SoC, DSS UNDERFLOW
occurs, the frequency of which increases for 720p60 and 1080p30.
Merging the FIFOs on all pipelines eliminated UNDERFLOW in most use cases.

This hack patch is added to provide interim solution for the TI release kernel
so that 720p use case over HDMI and camera and graphics team use cases work.

This patch adds this as a FIFO SIZE LIMITATION feat into DSS feature list
and enables the FIFO merge for this limitation.

Changes in V3:
the dss feat renamed to point to real issue - FIFO_MEM_BANDWIDTH_BUG
low threshold assignment made after checking for the feat

Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>
Signed-off-by: Sathya Prakash M R <sathyap@ti.com>